spack.package: add to docs, drop re-exports (#51006)
`spack.package` has the following aliases which are exported:

```python
env = os.environ
cd = os.chdir
pwd = os.getcwd
```

but it *also* re-exports `environ`, `chdir` and `getcwd`, which are
unused in builtin.

That's likely an oversight, and I don't consider it breaking to
remove those given the lack of use in builtin.

As a side-effect, we can now generate docs for spack.package, which was previously impossible due to formatting issues of the docstring of builtin posix.chdir.