spack diff: show runtime differences (#51076)
Currently, differences in `gcc-runtime` or `glibc` packages are hidden
by `spack diff`. This is an unintended artifact of how they are specially
handled in the solver.

This PR adds an option to `SpecSolverSetup.spec_clauses` to include
runtime specs that are not unified in the spec. The `spack diff` command
calls `spec_clauses` with the new option, to ensure those differences are
represented in the `spack diff` output.

Includes regression test.


---------

Signed-off-by: Gregory Becker <becker33@llnl.gov>
Co-authored-by: becker33 <becker33@users.noreply.github.com>