when context manager: deal with `%` vs. `^` correctly (#51259)
Fixes a bug where `with when("..."):` incorrectly relaxed `%dep` to `^dep`.

Removes the custom and faulty logic in `merge_abstract_anonymous_specs`, in favor of the existing and better tested `Spec.constrain` function, with the subtle different that virtuals are not resolved from configured package repositories. The constrain operation is entirely symbolic.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>